OpenGlider V0.1 Initial Release
This release was first made public for the security conference DEFCON
21, where @NickFlipper presented his talk "10000 Yen Into The Sea".
Some file extensions have been appended with .txt for the security of
conference attendees.

Because OpenGlider is intended to be an Open Source Hardware Project,
the solid models for "OpenGlider-V0.1"  were converted to a common
interchange format (IGES or Parasolid) & included in the conference DVD
provided to attendees .

About V0.1:
In order to maximize depth of operation, this release was designed to
rely entirely on the expansion of a phase change material(wax) in order
to change its displacement in the water. As a result, the only trapped
gasses in the vehicle were those contained within the glass microspheres
used in the Syntactic "Buoyancy Sand".

"Buoyancy Sand" is a new composite material where glass microspheres are
suspended in melted paraffin in order to achieve a malleable solid that
has a specific gravity well below 1.0.

Please Read!
Bench testing of this prototype design revealed a problem where the
selected coin cells composing the battery pack were a poor match for the
current requirements of the resistive heat source.

To resolve this issue, "OpenGlider-V0.2" uses a modified version of
"Nose_Body_Assemby_2of4.IGS" called
"Nose_Body_Assembly_Syringe_2of4.IGS" which allowed several low
cost(mass-produced) electric motors to drive 4x independent piston pumps
via an Acetal (POM) and FDM ABS plastic linear-actuator.

For further information please see the repository "10000 Yen" for a
video of the presentation and/or a copy of the slides.

Best Regards,
-Flipper
Email: nick@openglider.com
Twitter Handle: @NickFlipper